+++ hitchhiking_diary_server/components/images/router.py
@@ -1,0 +1,6 @@
+from fastapi import APIRouter
+
+from .endpoints import opengraph
+
+router = APIRouter()
+router.include_router(opengraph.router)