+++ hitchhiking_diary_server/components/rest/endpoints/tokens.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import Session
 from hitchhiking_diary_server.core import settings
 from hitchhiking_diary_server.db.session import get_db
 from hitchhiking_diary_server.models import User
-from hitchhiking_diary_server.rest.schemas.token import TokenFormSchema, TokenDetailSchema
+from hitchhiking_diary_server.components.rest.schemas.token import TokenFormSchema, TokenDetailSchema
 
 router = APIRouter()